orig_data/requirements_rf.txt
-f https://packages.rockfish.ai/index.html
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
asn1crypto==1.5.1
attrs==25.4.0
blis==1.3.3
boto3==1.42.51
botocore==1.42.51
catalogue==2.0.10
cattrs==25.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
cymem==2.0.13
datafusion==44.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.3.1
faker==19.10.0
filelock==3.24.2
fonttools==4.61.1
frozenlist==1.8.0
grpcio==1.78.0
idna==3.11
janus==1.2.0
jinja2==3.1.6
jmespath==1.1.0
kiwisolver==1.4.9
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
multidict==6.7.1
murmurhash==1.0.15
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pandas==2.3.3
phonenumbers==9.0.24
pillow==12.1.1
platformdirs==4.9.2
preshed==3.0.12
presidio-analyzer==2.2.361
propcache==0.4.1
protobuf==5.29.6
pyarrow==20.0.0
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyjwt==2.11.0
pyopenssl==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv
pytz==2025.2
pyyaml==6.0.3
regex==2026.1.15
requests==2.32.5
requests-file==3.0.1
rich==14.3.2
rockfish==0.53.0
s3transfer==0.16.0
scipy==1.17.0
seaborn==0.13.2
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smart-open==7.5.0
snowflake-connector-python==4.3.0
snowflake-sqlalchemy==1.8.2
sortedcontainers==2.4.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlalchemy==2.0.46
srsly==2.5.2
thinc==8.3.10
tldextract==5.3.1
tomlkit==0.14.0
tqdm==4.67.3
typer==0.24.0
typer-slim==0.24.0
types-requests==2.32.4.20260107
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
wrapt==2.1.1
yarl==1.22.0
